Rework a bit to make the code work more correctly.
While no data is yet communicated, this stubs out things well enough for
mosh-client to believe it is sending packets to nowhere, which causes it to
have the expected behavior of timing out. This, with some other changes, sets
up the basic structure for hooking up the communication with JS.

Notably, this required munging some original Mosh code, specifically the
randomness code of PRNG; this will need to be addressed.