Fix bug with recvfrom() emulation; bump version.
In UDP::Receive(), if there were no packets to read, it was returning 0
instead of -1 and EWOULDBLOCK. This caused Mosh to go crazy, because there is
a circumstance where it creates a lot of sockets and tries to read from all of
them, regardless of the result of pselect(). I didn't realize Mosh would
behave like that; regardless, the emulation of recvfrom() shouldn't be wrong.
This manifested as a crypto error, went into a spin, and rendered the session
useless.

Also fixed the NativeUDP::Received() callback to bail if it gets a negative
result.