Fix bug with x86_32 not starting up.
The issue seemed to be that pthread_create() will cause a call to getrlimit()
(but not on x86_64), which was stubbed out to avoid issues with Mosh trying to
turn off core dumps. The stub would cause pthread_create() to fail with EAGAIN.
Altered the stub to return something pthread_create() can swallow, while
keeping Mosh happy..

Also improved reporting in the event that the thread cannot be created.