Put all Mosh source files back to their original state.
This allows building with newlib without any modifications to the upstream
code. It uses a "magic" header file to modify the compiling environment of
Mosh. This file is in a directory which is first in the include path. assert.h
was chosen as it is included by most modules. This "magic" assert.h gets
included instead of the normal one, adds our special sauce, and includes the
usual one along the way